requirements-test.txt
# Minimal dependencies for running tests (no asyncpg, no WeasyPrint, no boto3)
pydantic==2.10.5
pydantic-settings==2.7.1
httpx==0.28.1
pyyaml==6.0.2
rich==13.9.4
pytest>=8.0
# Integration test deps (all have wheels, no C compilation)
fastapi==0.115.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pytest-asyncio>=0.23.0
python-multipart==0.0.20
jinja2==3.1.5
markdown==3.7
